Exception handling - delete

+++ src/main/java/com/brunopereira/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.brunopereira.course.entities.User;
 import com.brunopereira.course.repositories.UserRepository;
+import com.brunopereira.course.services.exceptions.DatabaseException;
+import com.brunopereira.course.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class UserService {
@@ -20,7 +23,7 @@ public class UserService {
 
   public User findById(Long id) {
     Optional<User> user = repository.findById(id);
-    return user.get();
+    return user.orElseThrow(() -> new ResourceNotFoundException(id));
   }
 
   public User insert(User user) {
@@ -28,6 +31,27 @@ public class UserService {
   }
 
   public void delete(Long id) {
-    repository.deleteById(id);
+    try {
+      if(repository.existsById(id)) {
+        repository.deleteById(id);
+      } else {
+        throw new ResourceNotFoundException(id);
+      }
+    } catch (DataIntegrityViolationException e) {
+      throw new DatabaseException(e.getMessage());
+    }
+  }
+
+  public User update(Long id, User user) {
+    User entity = repository.getReferenceById(id);
+    updateData(entity, user);
+
+    return repository.save(entity);
+  }
+
+  private void updateData(User entity, User user) {
+    entity.setName(user.getName());
+    entity.setEmail(user.getEmail());
+    entity.setPhone(user.getPhone());
   }
 }
